Added new fields and captions to EpiData
New fields added to the EpiDataDto and EpiData classes in the sormas-api module.
Added captions to caption files for newly added fields
Setter and getter methods have been implemented in the EpiDataFacadeEjb class to support the new fields.
UI modifications have been made in EpiDataForm for handling the new fields.